Separate MockControlClient with its tests
